vao is back in
Also removing the map referencing and instead will store the referenced
vbos in a set at time of setting attrib pointer. The downside is that I
need to say that getting the GLuint is const (it is for the sets needs)
but in reality the object could be completely changed from just the
GLuint, since the GLuint is the object really. So it's not actually
guaranteed const..